reorg project structure
