+++ src/helpers/widgetRegistration.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Todo from '../components/Todo/todo';
+import Todo from '../components/Todo';
 
 const Widgets = {
   key: <div> KEY ONE </div>,
