+++ src/popup.jsx
@@ -1,7 +1,7 @@
+/* global document */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Popup from './components/popup';
+import Popup from './components/Popup';
 import './index.css';
 
-/* global document */
 ReactDOM.render(<Popup />, document.getElementById('root'));
